refactor(sftp): extract SftpFile and related types into sftp_file.dart
